requirements.txt
anyio==3.7.0
asgiref==3.7.2
autohooks==23.4.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorful==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pontos==23.6.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rich==13.4.2
semver==3.0.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomlkit==0.11.8
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
